pallet_sudo::pallet

Trait DefaultConfig

pub trait DefaultConfig: DefaultConfig {
    type RuntimeEvent;
    type RuntimeCall;
    type WeightInfo: WeightInfo;
}

Based on Config. Auto-generated by #[pallet::config(with_default)]. Can be used in tandem with #[register_default_config] and #[derive_impl] to derive test config traits based on existing pallet config traits in a safe and developer-friendly way.

See here for more information and caveats about the auto-generated DefaultConfig trait and how it is generated.

Required Associated Types

type RuntimeEvent

The overarching event type.

type RuntimeCall

A sudo-able call.

type WeightInfo: WeightInfo

Type representing the weight of this pallet